Move ssh2pgp.c to kering_ssh.c
