+++ src/lib/rnp.c
@@ -86,7 +86,6 @@ __RCSID("$NetBSD: rnp.c,v 1.98 2016/06/28 16:34:40 christos Exp $");
 #include "rnpdefs.h"
 #include "crypto.h"
 #include "bn.h"
-#include "ssh2pgp.h"
 #include "defs.h"
 #include "../common/constants.h"
